Recompute header ranges when folding.
Previously header ranges could be broken to extend over previous lines.
The easiest way to reproduce is to put the caret at ^ in this line:

    fn ^foo() { … }

And insert a newline; the header will cover both the “fn “ and “foo() {
… }” lines.

This hacky patch recomputes the header range whenever a modification
happens inside it.